(pure-move) migrate encode/decode_tensor_data
