+++ ttnn/api/ttnn/tensor/tensor_impl.hpp
@@ -9,6 +9,7 @@
 
 #include <tt-metalium/mesh_device.hpp>
 #include <tt-metalium/tilize_utils.hpp>
+#include <tt-metalium/experimental/tensor/impl/tensor_impl.hpp>
 
 #include "ttnn/tensor/tensor.hpp"
 #include "ttnn/tensor/tensor_utils.hpp"
@@ -16,10 +17,6 @@
 #include "ttnn/tensor/layout/tensor_layout.hpp"
 
 namespace tt::tt_metal::tensor_impl {
-
-// Empty structs to facilitate Tensor template logic.
-struct bfloat4_b {};
-struct bfloat8_b {};
 
 // -----------------------------------------------------------------------------------------------------------------------------------------------
 // ===============================================================================================================================================
@@ -49,31 +46,6 @@ std::vector<T> convert_layout_tile_to_row_major(
         transpose_within_face,
         transpose_of_faces);
 }
-
-// Converts logical data into physical data based on tensor spec
-// - Logical data: Flat container of row major data corresponding to some ND logical shape
-// - Physical data: Flat container of physical data corresponding to tensor spec. It takes into account:
-//   * Sharding: Each shard will be padded to nearest page (if needed)
-//     ** This is mostly for logical sharding, since logical shards may not be aligned to page in general
-//     ** For interleaved, it will be handled as a "logically sharded" tensor with same shard shard height/width
-//        as the original tensor dims at -2 and -1. In the future, interleaved may be generalized as sharded.
-//     ** This means padding may be inserted in the middle of logical data (if needed)
-//   * Layout: Each aligned shard will be tilized (if needed)
-//     ** Tilization happens after first inserting padding to align shards (if needed)
-//     ** For the last shard, we only align to nearest page instead of full shard size for partial shards
-//   * After conversion, size of physical data will match 2D physical size indicated by tensor_spec.physical_shape()
-template <typename T>
-std::vector<T> encode_tensor_data(ttsl::Span<const T> logical_data, const TensorSpec& tensor_spec, T pad_value = 0);
-
-// Converts physical data into logical data based on tensor spec (see encode_tensor_data for details)
-// - Physical data: Flat container of physical data corresponding to tensor spec
-//   * Assumes that the physical data already matches tensor spec
-//   * There is a bare minimum check that size of physical data matches size indicated by tensor_spec.physical_shape()
-// - Logical data: Flat container of row major data corresponding to some ND logical shape
-//   * To get logical data, perform the exact inverse process of encode_tensor_data
-//   * Resulting data is safe to be converted to python tensors or general consumption with just a ND logical shape
-template <typename T>
-std::vector<T> decode_tensor_data(ttsl::Span<const T> physical_data, const TensorSpec& tensor_spec);
 
 // ===============================================================================================================================================
 //                                                              High Level APIs
@@ -250,30 +222,6 @@ std::string to_string(const Tensor& tensor);
 Tensor extract_shard(const Tensor& tensor, const uint32_t& core_id);
 
 HostTensor to_dtype(const HostTensor& input_tensor, DataType dtype);
-
-// Utility to convert runtime DataType to compile-time constant and dispatch the function call
-template <typename Func, typename... Args>
-auto dispatch(DataType dtype, Func&& func, Args&&... args) {
-    switch (dtype) {
-        case DataType::BFLOAT16:
-            return (std::forward<Func>(func)).template operator()<bfloat16>(std::forward<Args>(args)...);
-        case DataType::FLOAT32:
-            return (std::forward<Func>(func)).template operator()<float>(std::forward<Args>(args)...);
-        case DataType::INT32:
-            return (std::forward<Func>(func)).template operator()<int32_t>(std::forward<Args>(args)...);
-        case DataType::UINT32:
-            return (std::forward<Func>(func)).template operator()<uint32_t>(std::forward<Args>(args)...);
-        case DataType::UINT16:
-            return (std::forward<Func>(func)).template operator()<uint16_t>(std::forward<Args>(args)...);
-        case DataType::UINT8:
-            return (std::forward<Func>(func)).template operator()<uint8_t>(std::forward<Args>(args)...);
-        case DataType::BFLOAT8_B:
-            return (std::forward<Func>(func)).template operator()<bfloat8_b>(std::forward<Args>(args)...);
-        case DataType::BFLOAT4_B:
-            return (std::forward<Func>(func)).template operator()<bfloat4_b>(std::forward<Args>(args)...);
-        default: TT_THROW("Unsupported data type");
-    }
-}
 
 // ======================================================================================
 //                                  HostTensor Factory Functions
